==> methodology_blank/initial/1.1_First_Query_to_Twarc.ipynb <==
import pandas as pd
import pickle

**Manual input**: Write the path where you have the folder and the separator that your computer use.

path='C:/Users/acoub/OneDrive/Desktop/DSDM/Thesis/ChileGov/methodology_blank'
sep='/'

path_scripts = path+sep+'initial'
path_data= path+sep+'data'

**Manual input**: Put the list of words that are related with the topic of interest (visit web page semantic-link.com to obtain a first list and manaully filter by the most relevant)
  
Warning: The final list of words must have maximum 900 characteres aprox, so select only the relevants.

list_of_words=['word1','word2','word3']

**Manual input**: Select the period of time that you want to search. Format YYYY-MM-DD

start_date="YYYY-MM-DD"
end_date="YYYY-MM-DD"

filename = path_data+sep+'start_search_date.sav'
pickle.dump((start_date), open(filename, 'wb'))

filename = path_data+sep+'end_search_date.sav'
pickle.dump((end_date), open(filename, 'wb'))

Run this cell and follow the instructions of the output.

query=""
for word in list_of_words:
    query+= word + " OR "
query=query[:len(query)-4]

query="("+query+")"+ " -is:retweet place_country:CL"

if len(query)<=1024:
    print("Ready to run the next cell and search in twarc.")
else:
    print("Query too long, reduce the list of words.")

with open(path_data+sep+"first_search.txt", 'w') as f:
    f.write(query)

file=path_data+sep+"first_search.txt"

**Manual revision:** Run the following two cells to know the total number of tweets to download and see if does not exceed the limit of your twarc acount. If is OK run also the next cell.

%cd {path_data}
! twarc2 searches --archive --counts-only --granularity day --start-time {start_date} --end-time {end_date} {file} first_search_count.csv

count=pd.read_csv(path_data+sep+'first_search_count.csv',encoding = "ISO-8859-1")
print("Tweets to download: "+str(count['day_count'].sum()))

**Download the tweets:** (Only if you have enought limit in your twarc account)

%cd {path_data}
! twarc2 searches --archive --start-time {start_date} --end-time {end_date} {file} first_search.json
