==> methodology_blank/initial/2.1_Recovering_Hashtags_From_Politicians.ipynb <==
import json
from twarc_csv import DataFrameConverter
from twarc.expansions import ensure_flattened
import pandas as pd
from twarc_csv import dataframe_converter
from twarc_csv import CSVConverter, DataFrameConverter
import math
import emoji
from operator import itemgetter
import matplotlib.pyplot as plt
from datetime import datetime
from datetime import timedelta
from advertools import extract_emoji
import numpy as np
import pickle

**Manual Input**: Set the path.

path='C:/Users/acoub/OneDrive/Desktop/DSDM/Thesis/ChileGov/methodology_blank'
sep='/'

path_scripts = path+sep+'initial'
path_data= path+sep+'data'

**Manual Input**: List of politicians usernames to search the most popular hashtags.  
Alternative: Paste in the data folder one txt file named "Left_Politicians.txt" with one username per line for left politicians and the same for right politicians in a file "Right_Politicians.txt" 

left_wing=['rightusername1','rightusername2','rightusername3']
right_wing=['leftusername1','leftusername2','leftusername3']

If you added directly the txt file to the folder data, don't run the following cell.

with open(path_data+sep+'Left_Politicians.txt', 'w') as f:
    for line in left_wing:
        f.write(line)
        f.write('\n')
        
with open(path_data+sep+'Right_Politicians.txt', 'w') as f:
    for line in right_wing:
        f.write(line)
        f.write('\n')

**Manual input:** Define the period of time when you want to evaluate the use of hashtags to label Left and Right. We recomend a period when is easy to identificate right wing and left wing people (for instance an electoral period). Format YYYY-MM-DD

start_date="YYYY-MM-DD"
end_date="YYYY-MM-DD"

filename = path_data+sep+'start_date_for_label.sav'
pickle.dump((start_date), open(filename, 'wb'))

filename = path_data+sep+'end_date_for_label.sav'
pickle.dump((end_date), open(filename, 'wb'))

**Download the tweets**

file=path_data+sep+'Left_Politicians.txt'
%cd {path_data}
! twarc2 timelines --archive --start-time {start_date} --end-time {end_date} {file} left_hashtags.json

file=path_data+sep+'Right_Politicians.txt'
%cd {path_data}
! twarc2 timelines --archive --start-time {start_date} --end-time {end_date} {file} right_hashtags.json

Run the following cells only after you download the tweets.

converter = DataFrameConverter()
data = []
with open(path_data+sep+"left_hashtags.json") as f:
    for line in f:
        data.append(json.loads(line))
df_left = converter.process(data)

def hash_retrieve(df):
    """
    df : dataframe of tweets
    Description: 
        The function takes as an object a df of tweets obtained via twarc and 
        returns a generator object.
    
    """

    for line, id in zip(df['entities.hashtags'], df['id']):
        if pd.isna(line):
            continue
        line = line.strip()
        data = json.loads(line)
        for hashtag in ensure_flattened(data):
            yield [hashtag['tag'], id]

hashtags_df = pd.DataFrame(list(hash_retrieve(df_left)),
                 columns=['tweet_hashtags', 'id'])

hashtags_df = hashtags_df.groupby('id')['tweet_hashtags'].apply(lambda x: ','.join(x))

df_left = df_left.merge(hashtags_df, how='left', left_on='id', right_on='id')
df_left['tweet_hashtags']=df_left['tweet_hashtags'].str.lower()

hasht=[]
for i in df_left.index:
    try:
        hasht.append(df_left['tweet_hashtags'][i].split(","))
    except:
        hasht.append([])
    
    
df_left['list_hashtags']=hasht

hashtags=[]
for i in df_left.index:
    hashtags+=df_left['list_hashtags'][i]

**Manual revision:** Review the 100 most popular hashtags for left leaning politician during the selected period.

from collections import Counter
count = Counter(hashtags)
most_occur = count.most_common(100)
  
print(most_occur)

**Manual input** Decide the list of hashtags to label users. Think in hashtags that clearly identified the user with the left leaning. Don't use the # before the hashtag and use all the characters in lowercase. Max: 15 hashtags.

left_hashtags=["left_hashtag_1","left_hashtag_2","left_hashtag_3"]

converter = DataFrameConverter()
data = []
with open(path_data+sep+"right_hashtags.json") as f:
    for line in f:
        data.append(json.loads(line))
df_right = converter.process(data)

hashtags_df = pd.DataFrame(list(hash_retrieve(df_right)),
                 columns=['tweet_hashtags', 'id'])

hashtags_df = hashtags_df.groupby('id')['tweet_hashtags'].apply(lambda x: ','.join(x))

df_right = df_right.merge(hashtags_df, how='left', left_on='id', right_on='id')

df_right['tweet_hashtags']=df_right['tweet_hashtags'].str.lower()

hasht=[]
for i in df_right.index:
    try:
        hasht.append(df_right['tweet_hashtags'][i].split(","))
    except:
        hasht.append([])
    
    
df_right['list_hashtags']=hasht

hashtags=[]
for i in df_right.index:
    hashtags+=df_right['list_hashtags'][i]

**Manual revision:** Review the 100 most popular hashtags for right leaning politician during the selected period.

from collections import Counter
count = Counter(hashtags)
  
# most_common() produces k frequently encountered
# input values and their respective counts.
most_occur = count.most_common(100)
  
print(most_occur)

**Manual input** Decide the list of hashtags to label users. Think in hashtags that clearly identified the user with the right leaning. Don't use the # before the hashtag and use all the characters in lowercase. Max: 15 hashtags.

right_hashtags=["rigth_hashtag_1","right_hashtag_2","right_hashtag_3"]

filename = path_data+sep+'left_hashtags_for_label.sav'
pickle.dump((left_hashtags), open(filename, 'wb'))

filename = path_data+sep+'right_hashtags_for_label.sav'
pickle.dump((right_hashtags), open(filename, 'wb'))
